create protocol - in progress
going through this again to make sure I know where all the files are from and so I can capture images of all the steps. Currently waiting for all the .mzML files to load in Walnut in order to make the .elib file. 

I'm also doing this in a new folder in "grace" on the Desktop of Woodpecker, which will be uploaded to a file in /scaphapoda/grace on OWL so everyone can access all the files.